added functionality for infoboxes
This commit adds support for infobox generation. `infobox` is a new
optional parameter the user may specify. If passed to the `Map` class,
info boxes for each marker on the map are created and users may assign
a list of text/html strings for info box creation. If `infobox` is not
found in kwargs, the map just shows markers normally. I’ll provide an
example in the PR.

Also fixes some very minor PEP8 issues :)